+++ cms/src/layouts/RootLayout.tsx
@@ -1,12 +1,10 @@
-import React from "react";
 import { Outlet } from "@tanstack/react-router";
 
 import Header from "../components/header/Header";
 import Footer from "../components/footer/Footer";
-// import PrevLocationProvider from "../hooks/usePrevLocation";
 
-const DefaultLayout = () => {
-  console.log("DefaultLayout running..");
+const RootLayout = () => {
+  console.log("RootLayout running..");
 
   return (
     <>
@@ -19,4 +17,4 @@ const DefaultLayout = () => {
   );
 };
 
-export default DefaultLayout;
+export default RootLayout;