+++ cms/src/routes/__root.tsx
@@ -1,15 +1,12 @@
-import { createRootRouteWithContext } from "@tanstack/react-router";
-import DefaultLayout from "../layouts/DefaultLayout";
+import { createRootRouteWithContext, Outlet } from "@tanstack/react-router";
+
 import { AuthContext } from "../hooks/useAuth";
+import RootLayout from "../layouts/RootLayout";
 
 interface RouterContext {
-  auth: AuthContext | null;
+  auth: AuthContext;
 }
 
 export const Route = createRootRouteWithContext<RouterContext>()({
-  component: () => (
-    <>
-      <DefaultLayout />
-    </>
-  ),
+  component: RootLayout,
 });