+++ cms/src/routes/_dashboardLayout.tsx
@@ -1,0 +1,6 @@
+import { createFileRoute, Outlet, redirect } from "@tanstack/react-router";
+import DashboardLayout from "../layouts/DashboardLayout";
+
+export const Route = createFileRoute("/_protected")({
+  component: DashboardLayout,
+});